Updated the tutorial scripts as well as the path given to them in the info.py file. Replaced import of pylab with import of numpy and matplotlib in wafo.spectrum/models.py Replaced deprecated call "np.histogram(self.data, bins=limits, normed=True)" with "np.histogram(self.data, bins=limits, density=True)"

+++ src/wafo/objects.py
@@ -2003,7 +2003,7 @@ class TimeSeries(PlotData):
 
     def _interpolate(self, rate):
         if rate > 1:  # interpolate with spline
-            n = ceil(self.data.size * rate)
+            n = int(ceil(self.data.size * rate))
             ti = linspace(self.args[0], self.args[-1], n)
             x = stineman_interp(ti, self.args, self.data.ravel())
             # xi = interp1d(self.args, self.data.ravel(), kind='cubic')(ti)
